third_party/WebKit/LayoutTests/fast/dom/custom/created-callback.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
test(function () {
    var createdInvocations = 0;
    function created() {
        createdInvocations++;
    }

    var getterInvocations = 0;
    function getter() {
        getterInvocations++;
        return created;
    }

    function failer() {
        assert_unreached('the created callback must not be retrieved after registration');
    }

    var proto = Object.create(HTMLElement.prototype, {
      createdCallback: {
        get: getter
      }
    });
    var ctor = document.register('x-a', {prototype: proto});
    assert_equals(getterInvocations, 1, 'the created callback must have been retrieved');

    proto.createdCallback = failer;
    var element = new ctor();
    assert_equals(createdInvocations, 1, 'the created callback retrieved at registration must be invoked');
}, 'transfer created callback');
</script>
